// backend/src/common/enums.ts
export enum Role {
  ADMIN = 'ADMIN',
  MANAGER = 'MANAGER',
  MEMBER = 'MEMBER',
}

export enum Country {
  INDIA = 'INDIA',
  AMERICA = 'AMERICA',
}

export enum OrderStatus {
  PENDING = 'PENDING',
  PLACED = 'PLACED',
  CANCELLED = 'CANCELLED',
}

export enum PaymentType {
  CARD = 'CARD',
  UPI = 'UPI',
  WALLET = 'WALLET',
}